Adapt ilogtail

+++ ast/ddl.go
@@ -20,7 +20,6 @@ import (
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/mysql"
 	"github.com/pingcap/parser/terror"
-	"github.com/pingcap/parser/tidb"
 	"github.com/pingcap/parser/types"
 )
 
@@ -36,7 +35,6 @@ var (
 	_ DDLNode = &DropIndexStmt{}
 	_ DDLNode = &DropTableStmt{}
 	_ DDLNode = &DropSequenceStmt{}
-	_ DDLNode = &DropPlacementPolicyStmt{}
 	_ DDLNode = &RenameTableStmt{}
 	_ DDLNode = &TruncateTableStmt{}
 	_ DDLNode = &RepairTableStmt{}
@@ -503,9 +501,7 @@ func (n *ColumnOption) Restore(ctx *format.RestoreCtx) error {
 		pkTp := n.PrimaryKeyTp.String()
 		if len(pkTp) != 0 {
 			ctx.WritePlain(" ")
-			ctx.WriteWithSpecialComments(tidb.FeatureIDClusteredIndex, func() {
-				ctx.WriteKeyWord(pkTp)
-			})
+			ctx.WriteKeyWord(pkTp)
 		}
 	case ColumnOptionNotNull:
 		ctx.WriteKeyWord("NOT NULL")
@@ -578,12 +574,10 @@ func (n *ColumnOption) Restore(ctx *format.RestoreCtx) error {
 		ctx.WriteKeyWord("STORAGE ")
 		ctx.WriteKeyWord(n.StrValue)
 	case ColumnOptionAutoRandom:
-		ctx.WriteWithSpecialComments(tidb.FeatureIDAutoRandom, func() {
-			ctx.WriteKeyWord("AUTO_RANDOM")
-			if n.AutoRandomBitLength != types.UnspecifiedLength {
-				ctx.WritePlainf("(%d)", n.AutoRandomBitLength)
-			}
-		})
+		ctx.WriteKeyWord("AUTO_RANDOM")
+		if n.AutoRandomBitLength != types.UnspecifiedLength {
+			ctx.WritePlainf("(%d)", n.AutoRandomBitLength)
+		}
 	default:
 		return errors.New("An error occurred while splicing ColumnOption")
 	}
@@ -638,9 +632,7 @@ type IndexOption struct {
 func (n *IndexOption) Restore(ctx *format.RestoreCtx) error {
 	hasPrevOption := false
 	if n.PrimaryKeyTp != model.PrimaryKeyTypeDefault {
-		ctx.WriteWithSpecialComments(tidb.FeatureIDClusteredIndex, func() {
-			ctx.WriteKeyWord(n.PrimaryKeyTp.String())
-		})
+		ctx.WriteKeyWord(n.PrimaryKeyTp.String())
 		hasPrevOption = true
 	}
 	if n.KeyBlockSize > 0 {
@@ -941,44 +933,29 @@ func (n *ColumnDef) Validate() bool {
 	return !(generatedCol && illegalOpt4gc)
 }
 
-type TemporaryKeyword int
-
-const (
-	TemporaryNone TemporaryKeyword = iota
-	TemporaryGlobal
-	TemporaryLocal
-)
-
 // CreateTableStmt is a statement to create a table.
 // See https://dev.mysql.com/doc/refman/5.7/en/create-table.html
 type CreateTableStmt struct {
 	ddlNode
 
 	IfNotExists bool
-	TemporaryKeyword
-	// Meanless when TemporaryKeyword is not TemporaryGlobal.
-	// ON COMMIT DELETE ROWS => true
-	// ON COMMIT PRESERVE ROW => false
-	OnCommitDelete bool
-	Table          *TableName
-	ReferTable     *TableName
-	Cols           []*ColumnDef
-	Constraints    []*Constraint
-	Options        []*TableOption
-	Partition      *PartitionOptions
-	OnDuplicate    OnDuplicateKeyHandlingType
-	Select         ResultSetNode
+	IsTemporary bool
+	Table       *TableName
+	ReferTable  *TableName
+	Cols        []*ColumnDef
+	Constraints []*Constraint
+	Options     []*TableOption
+	Partition   *PartitionOptions
+	OnDuplicate OnDuplicateKeyHandlingType
+	Select      ResultSetNode
 }
 
 // Restore implements Node interface.
 func (n *CreateTableStmt) Restore(ctx *format.RestoreCtx) error {
-	switch n.TemporaryKeyword {
-	case TemporaryNone:
-		ctx.WriteKeyWord("CREATE TABLE ")
-	case TemporaryGlobal:
-		ctx.WriteKeyWord("CREATE GLOBAL TEMPORARY TABLE ")
-	case TemporaryLocal:
+	if n.IsTemporary {
 		ctx.WriteKeyWord("CREATE TEMPORARY TABLE ")
+	} else {
+		ctx.WriteKeyWord("CREATE TABLE ")
 	}
 	if n.IfNotExists {
 		ctx.WriteKeyWord("IF NOT EXISTS ")
@@ -1046,14 +1023,6 @@ func (n *CreateTableStmt) Restore(ctx *format.RestoreCtx) error {
 		}
 	}
 
-	if n.TemporaryKeyword == TemporaryGlobal {
-		if n.OnCommitDelete {
-			ctx.WriteKeyWord(" ON COMMIT DELETE ROWS")
-		} else {
-			ctx.WriteKeyWord(" ON COMMIT PRESERVE ROWS")
-		}
-	}
-
 	return nil
 }
 
@@ -1113,10 +1082,10 @@ func (n *CreateTableStmt) Accept(v Visitor) (Node, bool) {
 type DropTableStmt struct {
 	ddlNode
 
-	IfExists         bool
-	Tables           []*TableName
-	IsView           bool
-	TemporaryKeyword // make sense ONLY if/when IsView == false
+	IfExists    bool
+	Tables      []*TableName
+	IsView      bool
+	IsTemporary bool // make sense ONLY if/when IsView == false
 }
 
 // Restore implements Node interface.
@@ -1124,13 +1093,10 @@ func (n *DropTableStmt) Restore(ctx *format.RestoreCtx) error {
 	if n.IsView {
 		ctx.WriteKeyWord("DROP VIEW ")
 	} else {
-		switch n.TemporaryKeyword {
-		case TemporaryNone:
-			ctx.WriteKeyWord("DROP TABLE ")
-		case TemporaryGlobal:
-			ctx.WriteKeyWord("DROP GLOBAL TEMPORARY TABLE ")
-		case TemporaryLocal:
+		if n.IsTemporary {
 			ctx.WriteKeyWord("DROP TEMPORARY TABLE ")
+		} else {
+			ctx.WriteKeyWord("DROP TABLE ")
 		}
 	}
 	if n.IfExists {
@@ -1163,33 +1129,6 @@ func (n *DropTableStmt) Accept(v Visitor) (Node, bool) {
 		}
 		n.Tables[i] = node.(*TableName)
 	}
-	return v.Leave(n)
-}
-
-// DropPlacementPolicyStmt is a statement to drop a Policy.
-type DropPlacementPolicyStmt struct {
-	ddlNode
-
-	IfExists   bool
-	PolicyName model.CIStr
-}
-
-// Restore implements Restore interface.
-func (n *DropPlacementPolicyStmt) Restore(ctx *format.RestoreCtx) error {
-	ctx.WriteKeyWord("DROP PLACEMENT POLICY ")
-	if n.IfExists {
-		ctx.WriteKeyWord("IF EXISTS ")
-	}
-	ctx.WriteName(n.PolicyName.O)
-	return nil
-}
-
-func (n *DropPlacementPolicyStmt) Accept(v Visitor) (Node, bool) {
-	newNode, skipChildren := v.Enter(n)
-	if skipChildren {
-		return v.Leave(newNode)
-	}
-	n = newNode.(*DropPlacementPolicyStmt)
 	return v.Leave(n)
 }
 
@@ -1891,7 +1830,6 @@ type TableOption struct {
 	Default    bool
 	StrValue   string
 	UintValue  uint64
-	BoolValue  bool
 	TableNames []*TableName
 }
 
@@ -1925,33 +1863,17 @@ func (n *TableOption) Restore(ctx *format.RestoreCtx) error {
 		ctx.WritePlain("= ")
 		ctx.WriteKeyWord(n.StrValue)
 	case TableOptionAutoIncrement:
-		if n.BoolValue {
-			ctx.WriteWithSpecialComments(tidb.FeatureIDForceAutoInc, func() {
-				ctx.WriteKeyWord("FORCE")
-			})
-			ctx.WritePlain(" ")
-		}
 		ctx.WriteKeyWord("AUTO_INCREMENT ")
 		ctx.WritePlain("= ")
 		ctx.WritePlainf("%d", n.UintValue)
 	case TableOptionAutoIdCache:
-		ctx.WriteWithSpecialComments(tidb.FeatureIDAutoIDCache, func() {
-			ctx.WriteKeyWord("AUTO_ID_CACHE ")
-			ctx.WritePlain("= ")
-			ctx.WritePlainf("%d", n.UintValue)
-		})
+		ctx.WriteKeyWord("AUTO_ID_CACHE ")
+		ctx.WritePlain("= ")
+		ctx.WritePlainf("%d", n.UintValue)
 	case TableOptionAutoRandomBase:
-		if n.BoolValue {
-			ctx.WriteWithSpecialComments(tidb.FeatureIDForceAutoInc, func() {
-				ctx.WriteKeyWord("FORCE")
-			})
-			ctx.WritePlain(" ")
-		}
-		ctx.WriteWithSpecialComments(tidb.FeatureIDAutoRandomBase, func() {
-			ctx.WriteKeyWord("AUTO_RANDOM_BASE ")
-			ctx.WritePlain("= ")
-			ctx.WritePlainf("%d", n.UintValue)
-		})
+		ctx.WriteKeyWord("AUTO_RANDOM_BASE ")
+		ctx.WritePlain("= ")
+		ctx.WritePlainf("%d", n.UintValue)
 	case TableOptionComment:
 		ctx.WriteKeyWord("COMMENT ")
 		ctx.WritePlain("= ")
@@ -2042,15 +1964,11 @@ func (n *TableOption) Restore(ctx *format.RestoreCtx) error {
 			ctx.WritePlainf("%d", n.UintValue)
 		}
 	case TableOptionShardRowID:
-		ctx.WriteWithSpecialComments(tidb.FeatureIDTiDB, func() {
-			ctx.WriteKeyWord("SHARD_ROW_ID_BITS ")
-			ctx.WritePlainf("= %d", n.UintValue)
-		})
+		ctx.WriteKeyWord("SHARD_ROW_ID_BITS ")
+		ctx.WritePlainf("= %d", n.UintValue)
 	case TableOptionPreSplitRegion:
-		ctx.WriteWithSpecialComments(tidb.FeatureIDTiDB, func() {
-			ctx.WriteKeyWord("PRE_SPLIT_REGIONS ")
-			ctx.WritePlainf("= %d", n.UintValue)
-		})
+		ctx.WriteKeyWord("PRE_SPLIT_REGIONS ")
+		ctx.WritePlainf("= %d", n.UintValue)
 	case TableOptionPackKeys:
 		// TODO: not support
 		ctx.WriteKeyWord("PACK_KEYS ")
@@ -2262,7 +2180,6 @@ const (
 	AlterTableForce
 	AlterTableAddPartitions
 	AlterTableAlterPartition
-	AlterTablePartitionAttributes
 	AlterTableCoalescePartitions
 	AlterTableDropPartition
 	AlterTableTruncatePartition
@@ -2294,7 +2211,6 @@ const (
 	AlterTablePlacement
 	AlterTableAddStatistics
 	AlterTableDropStatistics
-	AlterTableAttributes
 )
 
 // LockType is the type for AlterTableSpec.
@@ -2394,7 +2310,6 @@ type AlterTableSpec struct {
 	PlacementSpecs  []*PlacementSpec
 	Writeable       bool
 	Statistics      *StatisticsSpec
-	AttributesSpec  *AttributesSpec
 }
 
 type TiFlashReplicaSpec struct {
@@ -2700,15 +2615,6 @@ func (n *AlterTableSpec) Restore(ctx *format.RestoreCtx) error {
 				return errors.Annotatef(err, "An error occurred while restore AlterTableSpec.PlacementSpecs[%d]", i)
 			}
 		}
-	case AlterTablePartitionAttributes:
-		ctx.WriteKeyWord("PARTITION ")
-		ctx.WriteName(n.PartitionNames[0].O)
-		ctx.WritePlain(" ")
-
-		spec := n.AttributesSpec
-		if err := spec.Restore(ctx); err != nil {
-			return errors.Annotatef(err, "An error occurred while restore AlterTableSpec.AttributesSpec")
-		}
 	case AlterTableCoalescePartitions:
 		ctx.WriteKeyWord("COALESCE PARTITION ")
 		if n.NoWriteToBinlog {
@@ -2908,12 +2814,6 @@ func (n *AlterTableSpec) Restore(ctx *format.RestoreCtx) error {
 				return errors.Annotatef(err, "An error occurred while restore AlterTableSpec.PlacementSpecs[%d]", i)
 			}
 		}
-	case AlterTableAttributes:
-		spec := n.AttributesSpec
-		if err := spec.Restore(ctx); err != nil {
-			return errors.Annotatef(err, "An error occurred while restore AlterTableSpec.AttributesSpec")
-		}
-
 	default:
 		// TODO: not support
 		ctx.WritePlainf(" /* AlterTableType(%d) is not supported */ ", n.Tp)
@@ -3364,13 +3264,6 @@ type PartitionMethod struct {
 
 	// Num is the number of (sub)partitions required by the method.
 	Num uint64
-
-	// KeyAlgorithm is the optional hash algorithm type for `PARTITION BY [LINEAR] KEY` syntax.
-	KeyAlgorithm *PartitionKeyAlgorithm
-}
-
-type PartitionKeyAlgorithm struct {
-	Type uint64
 }
 
 // Restore implements the Node interface
@@ -3379,11 +3272,6 @@ func (n *PartitionMethod) Restore(ctx *format.RestoreCtx) error {
 		ctx.WriteKeyWord("LINEAR ")
 	}
 	ctx.WriteKeyWord(n.Tp.String())
-
-	if n.KeyAlgorithm != nil {
-		ctx.WriteKeyWord(" ALGORITHM")
-		ctx.WritePlainf(" = %d", n.KeyAlgorithm.Type)
-	}
 
 	switch {
 	case n.Tp == model.PartitionTypeSystemTime:
@@ -3732,33 +3620,6 @@ func (n *PlacementSpec) Accept(v Visitor) (Node, bool) {
 		return v.Leave(newNode)
 	}
 	n = newNode.(*PlacementSpec)
-	return v.Leave(n)
-}
-
-type AttributesSpec struct {
-	node
-
-	Attributes string
-	Default    bool
-}
-
-func (n *AttributesSpec) Restore(ctx *format.RestoreCtx) error {
-	ctx.WriteKeyWord("ATTRIBUTES")
-	ctx.WritePlain("=")
-	if n.Default {
-		ctx.WriteKeyWord("DEFAULT")
-		return nil
-	}
-	ctx.WriteString(n.Attributes)
-	return nil
-}
-
-func (n *AttributesSpec) Accept(v Visitor) (Node, bool) {
-	newNode, skipChildren := v.Enter(n)
-	if skipChildren {
-		return v.Leave(newNode)
-	}
-	n = newNode.(*AttributesSpec)
 	return v.Leave(n)
 }
